Minor refactoring: removed unused method in KeyValueClient and moved SessionInfo in model.session package
