+++ src/main/java/com/orbitz/consul/SessionClient.java
@@ -1,7 +1,7 @@
 package com.orbitz.consul;
 
 import com.google.common.base.Optional;
-import com.orbitz.consul.model.kv.SessionInfo;
+import com.orbitz.consul.model.session.SessionInfo;
 import com.orbitz.consul.option.QueryOptions;
 import com.orbitz.consul.util.ClientUtil;
